Simple Video Tracking Software

Simple video tracking software with Shiny-based GUI.

trackR(tracker = "classic", ...)

Arguments

tracker
A character string indicating the tracking algorithm to use (see 'Details').

...
Arguments passed to runApp.

Value

This function does not return anything. A file is saved if the tracking is successfully completed.

Details

trackR provides three different tracking algorithms:

"classic":

This algorithm uses basic background subtraction to segment the objects in the image. A background image can be provided by the user or can be computed as the mean or median of n images taken at regular intervals throughout the video. "classic" should work well in most lab situations with a constant and homogenous background.

"quanta":

This algorithm uses quantized background subtraction to segment the objects in the image. In a quantized background image, each pixel corresponds to the lower (if the objects are generally darker than the background) or upper (if the objects are generally lighter than the background) 2.5 percentile of n images taken at regular intervals throughout the video. A pixel outside of these bounds in a given image would be considered as being part of an object. "quanta" should work better than "classic" in situations were the background is constant but not homogenous (e.g. if the background has different shades).

"tracktor":

This algorithm uses adaptive thresholding to segment the objects in the image. It is a liberal implementation of the tracktor algorithm developed for Python by Vivek Hari Sridhar, Dominique G. Roche and Simon Gingins. "tracktor" should work better than "classic" and "quanta" in situations were the background is not constant (e.g. when the environment changes over time). See "Sridhar, V. H., Roche, D. G., and Gingins, S. (2019). Tracktor: Image-based automated tracking of animal movement and behaviour. Methods Ecol. Evol. 10, 691. doi:10.1111/2041-210X.13166" for more details.

Examples

# Run trackR classic in an external window
if (FALSE) trackR("classic")

# Run trackR quanta in an external window
if (FALSE) trackR("quanta")

# Run trackR tracktor in an external window
if (FALSE) trackR("tracktor")

# Run trackR classic inside RStudio
if (FALSE) trackR("classic", launch.browser = rstudioapi::viewer)

# Run trackR quanta inside RStudio
if (FALSE) trackR("quanta", launch.browser = rstudioapi::viewer)

# Run trackR tracktor inside RStudio
if (FALSE) trackR("tracktor", launch.browser = rstudioapi::viewer)